requirements.txt
flask
pillow
numpy
tensorflow
scikit-learn
pandas
pickle
matplotlib
opencv-python
